% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roman.R
\docType{data}
\name{roman}
\alias{ROMAN}
\alias{roman}
\title{Roman numerals}
\format{An object of class \code{regex} (inherits from \code{character}) of length 1.}
\usage{
ROMAN

roman(lo, hi)
}
\arguments{
\item{lo}{A non-negative integer. Minimum number of repeats, when grouped.}

\item{hi}{positive integer. Maximum number of repeats, when grouped.}
}
\description{
Match roman numerals.
}
\examples{
# Constant form and character class
ROMAN
roman()

x <- c("MMMDCCCXLVIII", "MMMCMDCCCXLVIIV")
rx <- rebus.base::exactly(roman())
grepl(rx, x)
}
\seealso{
\code{\link[utils]{as.roman}}
}
\keyword{datasets}

